pypi2port: fix Python 3 support
The print() stuff is trivial.

For the removed encode() call: it was broken in Python 3 as
str('foo'.encode('utf-8')) gives a string "b'foo'", and the generated
Portfile contained such a line:

description     b'foo'

It's less than optimal. To fix the issue, just run string filtering with
the original (not-encoded) string. In Python 2, strings from xmlrpclib
is `unicode` if the returned data contains 8-bit characters (0x80 to
0xff), or simply `str` otherwise. In both cases, characters are correctly
filtered as Python 2 do unicode/str conversions under the hood. In
Python 3, both the summary string and `string.printable` are unicode, so
it works, too.

Tested with:
* python2.7 pypi2port.py -p structlog    # Only ASCII in the summary
* python3.6 pypi2port.py -p structlog
* python2.7 pypi2port.py -p python-lyric # Non-ASCII in the summary, see
* python3.6 pypi2port.py -p python-lyric # https://trac.macports.org/ticket/51932